back/requirements.txt
Flask==1.0.2
Flask-Session==0.3.0
Flask-SQLAlchemy==2.4.4
requests==2.22.0